refactor: extract HasSeats trait from BoardingPassBuilder
